requirements.txt
fuzzywuzzy==0.18.0
geopy==2.2.0
ipython==8.4.0
matplotlib==3.4.3
networkx==2.6.3
numpy==1.20.1
pandas==1.3.0
Pillow==9.1.1
plotly==5.3.1
scipy==1.7.1
setuptools==57.1.0
streamlit==1.9.2
pyhive[hive]
